# docs/openapi-mvp.yaml
openapi: 3.1.0
info:
  title: مقاول الشريك API
  version: 1.0.0
  description: Baseline MVP API. Detailed field rules live in SRS and API catalog.
servers:
- url: https://api.example.sa/api/v1
  description: Production placeholder
...
paths:
...
  /api/v1/auth/login:
    post:
      operationId: AUTH_002
      summary: تسجيل الدخول
      description: 'Authorization: Public. Rate limit: 10/min.'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericData'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden
        '409':
          description: State conflict
        '422':
          description: Business rule violation
        '429':
          description: Rate limited
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
            example:
              _fields: identifier,password
...
  /api/v1/auth/refresh:
    post:
      operationId: AUTH_004
      summary: تجديد الجلسة
      description: 'Authorization: Refresh token. Rate limit: 20/min.'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericData'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden
        '409':
          description: State conflict
        '422':
          description: Business rule violation
        '429':
          description: Rate limited
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
            example:
              _fields: refresh_token
...
  /api/v1/me:
    get:
      operationId: AUTH_006
      summary: بيانات المستخدم الحالي
      description: 'Authorization: Authenticated. Rate limit: 120/min.'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericData'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden
        '409':
          description: State conflict
        '422':
          description: Business rule violation
        '429':
          description: Rate limited
      security:
      - bearerAuth: []
...
  /api/v1/projects:
    post:
      operationId: PRJ_001
      summary: إنشاء مشروع
      description: 'Authorization: Contractor/Admin. Rate limit: 120/min.'
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericData'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden
        '409':
          description: State conflict
        '422':
          description: Business rule violation
        '429':
          description: Rate limited
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
            example:
              _fields: customer_id,name,type,location,value,dates,systems[]
    get:
      operationId: PRJ_002
      summary: قائمة المشاريع
      description: 'Authorization: Scoped. Rate limit: 120/min.'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericData'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden
        '409':
          description: State conflict
        '422':
          description: Business rule violation
        '429':
          description: Rate limited
      security:
      - bearerAuth: []
  /api/v1/projects/{id}:
    get:
      operationId: PRJ_003
      summary: تفاصيل المشروع
      description: 'Authorization: Project member/Admin. Rate limit: 120/min.'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericData'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden
        '409':
          description: State conflict
        '422':
          description: Business rule violation
        '429':
          description: Rate limited
      security:
      - bearerAuth: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    patch:
      operationId: PRJ_004
      summary: تحديث بيانات المشروع
      description: 'Authorization: Project owner/Admin. Rate limit: 120/min.'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericData'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden
        '409':
          description: State conflict
        '422':
          description: Business rule violation
        '429':
          description: Rate limited
      security:
      - bearerAuth: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
            example:
              _fields: partial project
...
  /api/v1/projects/{id}/documents:
    post:
      operationId: DOC_002
      summary: تسجيل مستند مرفوع
      description: 'Authorization: Project member. Rate limit: 120/min.'
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericData'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden
        '409':
          description: State conflict
        '422':
          description: Business rule violation
        '429':
          description: Rate limited
      security:
      - bearerAuth: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
            example:
              _fields: doc_type,title,storage_key,version,metadata
    get:
      operationId: DOC_003
      summary: مكتبة المستندات
      description: 'Authorization: Project member. Rate limit: 120/min.'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericData'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden
        '409':
          description: State conflict
        '422':
          description: Business rule violation
        '429':
          description: Rate limited
      security:
      - bearerAuth: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            request_id:
              type: string
    GenericData:
      type: object
      properties:
        data:
          type: object
          additionalProperties: true
        meta:
          type: object
          additionalProperties: true